Rearrange widgets and Streamline Data Colleciton
