+++ rfsocinterface/gui/loconfig.py
@@ -16,10 +16,14 @@ from rfsocinterface.core.settings import SettingsError
 from rfsocinterface.gui.uic.loconfig_ui import Ui_LoConfigWidget as Ui_LOConfigWidget
 from rfsocinterface.core.losweep import LoSweepData, LoSweep, DEFAULT_NCOLS
 from rfsocinterface.gui.lodiagnostics import DiagnosticsDialog
-from rfsocinterface.gui.utils import get_num_value
-from rfsocinterface.gui.widgets.progress_bar import IncrementalProgressDialog, QThreadJobProgressDialog, make_progress_dialog_incrementer
+from rfsocinterface.gui.widgets import (
+    get_num_value,
+    IncrementalProgressDialog,
+    make_progress_dialog_incrementer,
+    IconLabel,
+    ERROR_ICON_CODE
+)
 from rfsocinterface.core.rfsoc import RFSOCWrapper
-from rfsocinterface.gui.widgets.icon_label import IconLabel, ERROR_ICON_CODE
 from rfsocinterface.core.utils import ensure_path, get_filename, TabName
 from rfsocinterface.gui.main_widget import MainWidget
 
@@ -46,9 +50,6 @@ class LoConfigWidget(MainWidget, Ui_LOConfigWidget):
             or 'elevation'.
         tone_path (Path): The path to the selected tone list file.
     """
-    start_fit = Signal(object, QThreadJobProgressDialog, object, object, int, bool)
-    start_plot = Signal(object, DiagnosticsDialog, QThreadJobProgressDialog)
-
     def __init__(self, main_window: 'MainWindow', rfsocs: list[RFSOCWrapper], settings: dict, parent: QWidget | None=None) -> None:
         """Initialize the LO configuration window."""
         super().__init__(main_window, rfsocs, settings, parent=parent)
@@ -56,9 +57,6 @@ class LoConfigWidget(MainWidget, Ui_LOConfigWidget):
 
         self._sweep_dialog_results = []
         self._second_sweep_ran = False
-
-        self.start_fit.connect(self._save_and_fit_sweep)
-        self.start_plot.connect(self._plot_fit)
 
         self.channel_comboBox.set_default_title('Select Channels...')
 
@@ -507,188 +505,6 @@ class LoConfigWidget(MainWidget, Ui_LOConfigWidget):
         freqs, depths = finder.find_resonators()
         pdb.set_trace()
 
-
-
-
-    def run_sweep(self, rfsoc: RFSOCWrapper, chan: int):
-        chan_name = rfsoc.get_channel_name(chan)
-
-        # Get values from GUI, converting KHz to Hz
-        tone_shift = get_num_value(self.global_shift_lineEdit) * 1e3
-        diff_to_flag = get_num_value(self.flagging_lineEdit, float) * 1e3
-        freq_step = get_num_value(self.df_lineEdit)  * 1e3
-        full_span = get_num_value(self.deltaf_lineEdit)  * 1e3
-
-        savefile = get_filename(
-            file_type="LO", chan_name=chan_name, mkdir=True
-        )
-        match self.buttonGroup.checkedButton():
-            case self.filename_elevation_radioButton:
-                savefile = savefile.with_stem(f'{savefile.stem}_elev_{self.filename_elevation_lineEdit.text()}')
-            case self.filename_temperature_radioButton:
-                savefile = savefile.with_stem(f'{savefile.stem}_temp_{self.filename_temperature_lineEdit.text()}')
-            case _:
-                pass
-
-        pd = QProgressDialog(
-            'Setting Up LO Sweep...',
-            'Cancel',
-            0,
-            100,
-            parent=self,
-        )
-        pd.setAutoClose(True)
-        pd.show()
-        QApplication.processEvents()
-        
-        # For running on ONR compupter
-        sweep = LoSweep(
-            rfsoc,
-            chan,
-            savefile,
-            tone_shift,
-            freq_step,
-            full_span,
-            diff_to_flag=diff_to_flag,
-        )
-        pd.canceled.connect(sweep.cancel)
-        pd.setValue(0)
-        pd.setMinimum(0)
-        pd.setMaximum(sweep.n_steps)
-
-        def increment_progress():
-            nonlocal pd
-            pd.setValue(pd.value() + 1)
-
-        pd.setLabelText('Running LO Sweep...')
-        sweep_thread = Thread(target=sweep.run_sweep, args=(increment_progress,))
-        sweep_thread.start()
-
-        while not (sweep._processed or sweep._cancel):
-            QApplication.processEvents()
-            time.sleep(0.1)
-        
-        sweep_thread.join()
-        if sweep._cancel:
-            _logger.info('LO Sweep Cancelled')
-            return
-
-        # TODO
-        # Fit sweep if requested...
-        #    Multiprocessing???
-        # Show diagnostics dialog if requested...
-        sweep_data = sweep.data
-        sweep_data.set_diff_to_flag(get_num_value(self.flagging_lineEdit, float) * 1e3)
-        self.save_sweep(savefile, sweep_data)
-
-        pd = QThreadJobProgressDialog(
-            labelText='Fitting LO Sweep...',
-            maximum=sweep_data.ngoodchan,
-            parent=self,
-        )
-        pd.setAutoClose(False)
-        self._save_and_fit_sweep(sweep, pd, savefile, rfsoc, chan, False)
-
-
-    @ensure_path(3)
-    def _save_and_fit_sweep(self, sweep: LoSweep, pd: QThreadJobProgressDialog, savefile: Path, rfsoc: RFSOCWrapper, chan: int, second_sweep: bool=False):
-        sweep_data = self._wait_and_save(sweep, savefile, rfsoc, chan)
-        
-        # Make diagnostics window and setup connections
-        dw = DiagnosticsDialog(sweep_data, savefile, parent=self)
-        dw.finished.connect(lambda result: self._finish_sweep(result, savefile, sweep_data, rfsoc, chan, dw, second_sweep))
-        dw.upload_pushButton.clicked.connect(lambda: self._write_new_tones(sweep_data, rfsoc, chan))
-
-        pd.setValue(0)
-        pd.setLabelText('Fitting sweep results...')
-        pd.setMaximum(sweep_data.ngoodchan)
-        QApplication.processEvents()
-        pd.make_pool()
-        future = sweep_data.fit(pd=pd)
-        future.add_done_callback(lambda _: self.start_plot.emit(sweep_data, dw, pd))
-    
-    def _plot_fit(self, sweep_data: LoSweepData, dw: DiagnosticsDialog, pd: QThreadJobProgressDialog):
-        pd.setValue(0)
-        pd.setLabelText('Plotting fit results...')
-        pd.setMaximum(sweep_data.nchan)
-        pd.setAutoClose(True)
-        QApplication.processEvents()
-        fig, future = dw.plot(pd=pd)
-        dw.set_figure(fig)
-        future.add_done_callback(lambda _: fig.tight_layout())
-        future.add_done_callback(lambda _: dw.update_median_shift())
-        future.add_done_callback(lambda _: dw.show())
-    
-    @ensure_path(2)
-    def _finish_sweep(self, result: int, savefile: Path, sweep_data: LoSweepData, rfsoc: RFSOCWrapper, chan: int, dw: DiagnosticsDialog, second_sweep: bool=False):
-        # If the sweep was discarded, close without saving changes
-        if result == QDialog.DialogCode.Rejected:
-            _logger.debug(f'Diagnostics dialog rejected.')
-            return
-        self.save_sweep(savefile, sweep_data)
-        if not second_sweep:
-            if self.save_plots_CheckBox.isChecked():
-                _logger.debug(f'Saving LO sweep plots')
-                dw.save_plots()
-                plt.close('all')
-            if self.second_sweep_checkBox.isChecked():
-                self._write_new_tones(sweep_data, rfsoc, chan)
-                self.second_sweep(savefile, rfsoc, chan)
-            elif self.upload_checkBox.isChecked():
-                self._write_new_tones(sweep_data, rfsoc, chan)
-        else:
-            if self.second_sweep_save_plots_checkBox.isChecked():
-                _logger.debug(f'Saving LO sweep plots')
-                dw.save_plots()
-                plt.close('all')
-            
-    @ensure_path(1)
-    def second_sweep(self, first_sweep_savefile: Path, rfsoc: RFSOCWrapper, chan: int):
-        """Perform second LO sweep."""
-        _logger.debug(f'Performing second LO sweep...')
-        filename = first_sweep_savefile.stem + '_high_res.h5'
-        savefile = first_sweep_savefile.with_name(filename)
-
-        valon = rfsoc.get_valon(chan)
-
-        # For running on ONR Computer
-        lo_freq = rfsoc.get_channel(chan).lo_freq
-        rfsoc.set_frequency(chan, lo_freq)
-
-        rfchan = rfsoc.get_channel(chan)
-        tone_list = rfsoc.get_tone_list(chan)[0]
-        sweep = LoSweep(
-            valon,
-            rfchan,
-            tone_list,
-            lo_freq,
-        )
-        freq_step = get_num_value(self.second_sweep_df_lineEdit)  * 1e3  # KHz to Hz
-        full_span = get_num_value(self.deltaf_lineEdit)  * 1e3 / 5  # KHz to Hz
-        n_steps = full_span / freq_step
-
-        pd = QThreadJobProgressDialog(labelText='Running Second LO Sweep...',  maximum=n_steps, max_workers=1, parent=self)
-        pd.setAutoClose(False)
-        pd.show()
-        QApplication.processEvents()
-
-        sweep_data_future = sweep.run_sweep(rfsoc.get_channel(chan).chanmask, tone_list, N_steps=n_steps, freq_step=freq_step, pd=pd)
-        sweep_data_future.add_done_callback(lambda _: self.start_fit.emit(sweep, pd, savefile, rfsoc, chan, True))
-
-    def _wait_and_save(self, sweep: LoSweep, savefile: Path, rfsoc: RFSOCWrapper, chan: int):
-        counter = 0
-        while not sweep._processed:
-            counter += 1
-            if counter % 500 == 0:
-                _logger.debug('Waiting for sweep to finish...')
-            QApplication.processEvents()
-            time.sleep(0.1)
-        _logger.debug(f'LoConfigWidget finished waiting for LO Sweep processing. Saving to {str(savefile)}.')
-        sweep_data = sweep.data
-        sweep_data.set_diff_to_flag(get_num_value(self.flagging_lineEdit, float) * 1e3)
-        self.save_sweep(savefile, sweep_data)
-        return sweep_data
-    
     def _write_new_tones(self, sweep_data: LoSweepData, rfsoc: RFSOCWrapper, chan: int):
         """Write the new tones from fitting an LO Sweep to the RFSoC"""
         tone_file = get_filename(file_type='tonelist', chan_name=rfsoc.get_channel_name(chan))
